+++ packages/gatsby-plugin-typography/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.3.10...gatsby-plugin-typography@2.3.11) (2019-10-04)
+
+**Note:** Version bump only for package gatsby-plugin-typography
+
 ## [2.3.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-typography@2.3.8...gatsby-plugin-typography@2.3.10) (2019-09-26)
 
 **Note:** Version bump only for package gatsby-plugin-typography