+++ packages/gatsby-source-npm-package-search/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@2.1.12...gatsby-source-npm-package-search@2.1.13) (2019-10-04)
+
+### Bug Fixes
+
+- update dependency algoliasearch to ^3.35.0 ([#17910](https://github.com/gatsbyjs/gatsby/issues/17910)) ([436f11d](https://github.com/gatsbyjs/gatsby/commit/436f11d))
+
 ## [2.1.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@2.1.10...gatsby-source-npm-package-search@2.1.12) (2019-09-26)
 
 **Note:** Version bump only for package gatsby-source-npm-package-search
@@ -51,7 +57,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Bug Fixes
 
-- **www:** fallback to github & unpkg to get README.md ([#16322](https://github.com/gatsbyjs/gatsby/issues/16322)) ([f38b40d](https://github.com/gatsbyjs/gatsby/commit/f38b40d))
+- **www:** fallback to GitHub & unpkg to get README.md ([#16322](https://github.com/gatsbyjs/gatsby/issues/16322)) ([f38b40d](https://github.com/gatsbyjs/gatsby/commit/f38b40d))
 
 ## [2.1.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@2.1.1...gatsby-source-npm-package-search@2.1.2) (2019-07-12)