+++ packages/gatsby-source-shopify/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.0.20...gatsby-source-shopify@3.0.21) (2019-10-08)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
+## [3.0.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.0.19...gatsby-source-shopify@3.0.20) (2019-10-04)
+
+**Note:** Version bump only for package gatsby-source-shopify
+
 ## [3.0.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-shopify@3.0.17...gatsby-source-shopify@3.0.19) (2019-09-26)
 
 **Note:** Version bump only for package gatsby-source-shopify